log: Create gRPC client and send logs
This creates a gRPC client to the log service. The gRPC client is
declared as an exported variable to allow unit testing. It was exported
instead of using dependency injection because doing the latter would
make very confusing why the methods in the middle of the chain needs the
gRPC client as argument.

Signed-off-by: Getulio Valentin Sánchez <valentin2507@gmail.com>